Refactor: split UsageDashboard into separate component files
- Extract RequestUsageBar into its own file for reusability and
  single-responsibility (follows CLAUDE.md one-component-per-file rule)
- Extract AgentBreakdownTable into its own file with clear props interface
- Remove non-null assertions (!) in UsageDashboard — use proper type
  narrowing by extracting byAgent into a local variable before the null
  check, so TypeScript can narrow the type without assertions
- UsageDashboard.tsx reduced from 325 to 170 lines
