+++ frontend/src/pages/billing/UsageDashboard.tsx
@@ -1,5 +1,3 @@
-import { useMemo } from "react";
-import { Link } from "react-router-dom";
 import { Activity, Clock, Loader2, MessageSquare } from "lucide-react";
 import {
   Card,
@@ -8,18 +6,11 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
 import { useBillingUsage } from "@/hooks/useBillingUsage";
-import { useAgents } from "@/hooks/useAgents";
 import type { Plan, Subscription } from "@/hooks/useBillingPlan";
 import { formatCents, formatDate } from "./formatters";
+import { RequestUsageBar } from "./RequestUsageBar";
+import { AgentBreakdownTable } from "./AgentBreakdownTable";
 
 interface UsageDashboardProps {
   plan: Plan;
@@ -31,144 +22,6 @@ function daysRemaining(periodEnd: string): number {
   const now = new Date();
   const diff = end.getTime() - now.getTime();
   return Math.max(0, Math.ceil(diff / (1000 * 60 * 60 * 24)));
-}
-
-function RequestUsageBar({
-  total,
-  limit,
-}: {
-  total: number;
-  limit: number | null;
-}) {
-  const isUnlimited = limit === null;
-  const percentage = isUnlimited
-    ? 0
-    : limit > 0
-      ? Math.min((total / limit) * 100, 100)
-      : 0;
-  const isNearLimit = !isUnlimited && limit > 0 && percentage >= 80;
-  const isAtLimit = !isUnlimited && limit > 0 && total >= limit;
-
-  return (
-    <div className="space-y-1.5">
-      <div className="flex items-center justify-between text-sm">
-        <span className="font-medium">Requests this period</span>
-        <span className="text-muted-foreground tabular-nums">
-          {total.toLocaleString()} /{" "}
-          {isUnlimited ? "Unlimited" : limit.toLocaleString()}
-        </span>
-      </div>
-      {!isUnlimited && (
-        <div
-          className="h-2.5 w-full overflow-hidden rounded-full bg-muted"
-          role="progressbar"
-          aria-valuenow={Math.min(total, limit)}
-          aria-valuemin={0}
-          aria-valuemax={limit}
-          aria-label={`Requests: ${total.toLocaleString()} of ${limit.toLocaleString()} used`}
-        >
-          <div
-            className={`h-full rounded-full transition-all ${
-              isAtLimit
-                ? "bg-destructive"
-                : isNearLimit
-                  ? "bg-amber-500"
-                  : "bg-primary"
-            }`}
-            style={{ width: `${percentage}%` }}
-          />
-        </div>
-      )}
-    </div>
-  );
-}
-
-interface AgentBreakdownTableProps {
-  byAgent: Record<string, number>;
-  totalRequests: number;
-  showCost: boolean;
-  costCents: number;
-}
-
-function AgentBreakdownTable({
-  byAgent,
-  totalRequests,
-  showCost,
-  costCents,
-}: AgentBreakdownTableProps) {
-  const { agents } = useAgents();
-
-  const agentNameMap = useMemo(() => {
-    const map = new Map<string, string>();
-    for (const agent of agents) {
-      const meta = agent.metadata as Record<string, unknown> | undefined;
-      const name =
-        typeof meta?.name === "string" ? meta.name : `Agent ${agent.agent_id}`;
-      map.set(String(agent.agent_id), name);
-    }
-    return map;
-  }, [agents]);
-
-  const sorted = Object.entries(byAgent).sort(([, a], [, b]) => b - a);
-  const maxCount = sorted.length > 0 ? (sorted[0]?.[1] ?? 0) : 0;
-
-  return (
-    <Table>
-      <TableHeader>
-        <TableRow>
-          <TableHead>Agent</TableHead>
-          <TableHead className="text-right">Requests</TableHead>
-          <TableHead className="text-right">% of total</TableHead>
-          {showCost && <TableHead className="text-right">Est. cost</TableHead>}
-        </TableRow>
-      </TableHeader>
-      <TableBody>
-        {sorted.map(([agentId, count]) => {
-          const pctNum =
-            totalRequests > 0 ? (count / totalRequests) * 100 : 0;
-          const pct = pctNum.toFixed(1);
-          const barWidth =
-            maxCount > 0 ? Math.max(2, (count / maxCount) * 100) : 0;
-          const agentCost =
-            totalRequests > 0
-              ? Math.round((count / totalRequests) * costCents)
-              : 0;
-
-          return (
-            <TableRow key={agentId}>
-              <TableCell className="font-medium">
-                <div className="space-y-1">
-                  <Link
-                    to={`/agents/${agentId}`}
-                    className="text-primary hover:underline"
-                  >
-                    {agentNameMap.get(agentId) ?? `Agent ${agentId}`}
-                  </Link>
-                  <div className="h-1 w-full max-w-[120px] overflow-hidden rounded-full bg-muted">
-                    <div
-                      className="h-full rounded-full bg-primary/40"
-                      style={{ width: `${barWidth}%` }}
-                    />
-                  </div>
-                </div>
-              </TableCell>
-              <TableCell className="text-right tabular-nums">
-                {count.toLocaleString()}
-              </TableCell>
-              <TableCell className="text-right tabular-nums text-muted-foreground">
-                {pct}%
-              </TableCell>
-              {showCost && (
-                <TableCell className="text-right tabular-nums text-muted-foreground">
-                  {formatCents(agentCost)}
-                </TableCell>
-              )}
-            </TableRow>
-          );
-        })}
-      </TableBody>
-    </Table>
-  );
 }
 
 export function UsageDashboard({ plan, subscription }: UsageDashboardProps) {
@@ -213,10 +66,8 @@ export function UsageDashboard({ plan, subscription }: UsageDashboardProps) {
     );
   }
 
-  const hasBreakdown =
-    usage.breakdown?.by_agent &&
-    Object.keys(usage.breakdown.by_agent).length > 0;
-
+  const byAgent = usage.breakdown?.by_agent;
+  const hasBreakdown = byAgent != null && Object.keys(byAgent).length > 0;
   const totalCostCents = usage.requests.cost_cents + usage.sms.cost_cents;
 
   return (
@@ -287,7 +138,7 @@ export function UsageDashboard({ plan, subscription }: UsageDashboardProps) {
             <div className="space-y-2">
               <h3 className="text-sm font-medium">Usage by Agent</h3>
               <AgentBreakdownTable
-                byAgent={usage.breakdown!.by_agent!}
+                byAgent={byAgent}
                 totalRequests={usage.requests.total}
                 showCost={!isFree}
                 costCents={totalCostCents}
